+++ src/Core/MedicalCenters.Application/Features/MedicalWard/Queries/MedicalWard.cs
@@ -19,7 +19,7 @@ namespace MedicalCenters.Application.Features.MedicalWard.Queries
         {
             var response = new BaseQueryResponse();
             cancellationToken.ThrowIfCancellationRequested();
-            var result = await medicalWardRepository.Get((int)request.Id, cancellationToken);
+            var result = await medicalWardRepository.Get(request.Id, cancellationToken);
             if (result == null)
             {
                 throw new NotFoundException(Domain.Entities.MedicalWard.EntityTitle, request.Id.ToString());
@@ -36,6 +36,6 @@ namespace MedicalCenters.Application.Features.MedicalWard.Queries
 
     public record MedicalWardQuery : IRequest<BaseQueryResponse>
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
     }
 }
